use getENSNFTAvatarUrl
