+++ src/hooks/useENSPendingRegistrations.tsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import { useAccountSettings, useENSRegistration } from '.';
 import { ENSRegistrationState } from '@rainbow-me/entities';
 import { AppState } from '@rainbow-me/redux/store';
-import { isENSNFTAvatar, parseENSNFTAvatar } from '@rainbow-me/utils';
+import getENSNFTAvatarUrl from '@rainbow-me/utils/getENSNFTAvatarUrl';
 
 export default function useENSPendingRegistrations() {
   const { accountAddress } = useAccountSettings();
@@ -38,29 +38,7 @@ export default function useENSPendingRegistrations() {
   const registrationImages = useMemo(() => {
     const registrationImagesArray = pendingRegistrations.map(
       ({ name, records }) => {
-        let avatarUrl;
-        if (records?.avatar) {
-          const isNFTAvatar = isENSNFTAvatar(records.avatar);
-          if (isNFTAvatar) {
-            const { contractAddress, tokenId } = parseENSNFTAvatar(
-              records?.avatar
-            );
-            const uniqueToken = uniqueTokens.find(
-              token =>
-                token.asset_contract.address === contractAddress &&
-                token.id === tokenId
-            );
-            if (uniqueToken?.image_thumbnail_url) {
-              avatarUrl = uniqueToken?.image_thumbnail_url;
-            }
-          } else if (
-            records.avatar.startsWith('http') ||
-            (records.avatar.startsWith('/') &&
-              !records.avatar.match(/^\/(ipfs|ipns)/))
-          ) {
-            avatarUrl = records.avatar;
-          }
-        }
+        const avatarUrl = getENSNFTAvatarUrl(uniqueTokens, records?.avatar);
         return {
           avatarUrl,
           name,
